% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbarthetadist.R
\name{xbarthetadist}
\alias{xbarthetadist}
\title{Distribution plots}
\usage{
xbarthetadist(n, iter, mu, sigma, ci, con)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations}

\item{mu}{population mean}

\item{sigma}{population variance-covariance matrix}

\item{ci}{confidence}

\item{con}{coefficients for linear combination}
}
\value{

}
\description{
Creates the ggplot variables and puts them into a shiny dashboard in a plot that has already been
created by another function
}
